defects/pyproject.toml
[tool.poetry]
name = "defects-new"
version = "1.0.0"
description = ""
authors = ["danil <emurashin@gmail.com>"]
readme = "README.md"
packages = [{include = "defects_new"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
sqlmodel = "^0.0.8"
alembic = "^1.11.1"
asyncpg = "^0.27.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.9"}
fastapi-pagination = "^0.12.4"
fastapi-async-sqlalchemy = "0.3.12"
loguru = "^0.6.0"
aiosql = "^6.1"
aio-pika = "^9.3.0"
prometheus-fastapi-instrumentator = "^6.1.0"
ruff = "^0.1.9"


[tool.ruff]
line-length = 120
show-fixes = true
src = ["app"]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
]


[tool.ruff.per-file-ignores]
"app/models/string_defect_types.py" = ["F821"]
"app/models/tower_defect_types.py" = ["F821"]
"app/models/wire_defect_types.py" = ["F821"]
"app/models/track_defect_types.py" = ["F821"]
"app/models/__init__.py" = ["F401"]
"app/crud/__init__.py" = ["F401"]


[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

